+++ client/src/ts/modules/hangout/suggestions/hangoutSuggestions.ts
@@ -4,7 +4,7 @@ import { HANGOUT_SUGGESTIONS_LIMIT, HANGOUT_VOTING_STAGE } from "../../global/cl
 import { createDivElement, createParagraphElement } from "../../global/domUtils";
 import LoadingModal from "../../global/LoadingModal";
 import popup from "../../global/popup";
-import { getHangoutSuggestionsService } from "../../services/suggestionsServices";
+import { addHangoutSuggestionLikeService, getHangoutSuggestionsService } from "../../services/suggestionsServices";
 import { hangoutAvailabilityState, initHangoutAvailability } from "../availability/hangoutAvailability";
 import { globalHangoutState } from "../globalHangoutState";
 import { Suggestion } from "../hangoutTypes";
@@ -20,6 +20,8 @@ interface HangoutSuggestionsState {
 
   pageCount: number,
   pageItemsCount: number,
+
+  sortingMode: 'likes' | 'votes',
 };
 
 export const hangoutSuggestionState: HangoutSuggestionsState = {
@@ -31,6 +33,8 @@ export const hangoutSuggestionState: HangoutSuggestionsState = {
 
   pageCount: 1,
   pageItemsCount: 0,
+
+  sortingMode: 'likes',
 };
 
 const suggestionsRemainingSpan: HTMLSpanElement | null = document.querySelector('#suggestions-remaining-span');
@@ -73,6 +77,8 @@ function loadEventListeners(): void {
 
     initHangoutSuggestions();
   });
+
+  suggestionsContainer?.addEventListener('click', handleSuggestionsContainerClicks);
 };
 
 async function getHangoutSuggestions(): Promise<void> {
@@ -182,4 +188,160 @@ function updateRemainingSuggestionsCount(): void {
 
   const suggestionsCount: number = HANGOUT_SUGGESTIONS_LIMIT - globalHangoutState.data.suggestionsCount;
   suggestionsRemainingSpan.textContent = `${suggestionsCount}`;
+};
+
+async function handleSuggestionsContainerClicks(e: MouseEvent): Promise<void> {
+  if (!(e.target instanceof HTMLButtonElement)) {
+    return;
+  };
+
+  const suggestionElement: Element | null = e.target.closest('.suggestion');
+
+  if (!suggestionElement || !(suggestionElement instanceof HTMLDivElement)) {
+    return;
+  };
+
+  const suggestionIdString: string | null = suggestionElement.getAttribute('data-suggestionId');
+
+  if (!suggestionIdString || !Number.isInteger(+suggestionIdString)) {
+    return;
+  };
+
+  const suggestionId: number = +suggestionIdString;
+
+  if (e.target.classList.contains('view-suggestion-btn')) {
+    suggestionElement.classList.toggle('expanded');
+    return;
+  };
+
+  if (e.target.classList.contains('like-suggestion-btn')) {
+    if (suggestionElement.classList.contains('like-pending')) {
+      return;
+    };
+
+    if (suggestionElement.classList.contains('liked')) {
+      // TODO: implement unlike logic
+      return;
+    };
+
+    await addHangoutSuggestionLike(suggestionId, suggestionElement);
+  };
+};
+
+async function addHangoutSuggestionLike(suggestionId: number, suggestionElement: HTMLDivElement): Promise<void> {
+  if (!globalHangoutState.data) {
+    popup('Failed to like suggestion.', 'error');
+    return;
+  };
+
+  if (hangoutSuggestionState.memberLikesSet.has(suggestionId)) {
+    return;
+  };
+
+  suggestionElement.classList.add('like-pending');
+  const { hangoutMemberId, hangoutId } = globalHangoutState.data;
+
+  try {
+    await addHangoutSuggestionLikeService({ suggestionId, hangoutMemberId, hangoutId });
+    hangoutSuggestionState.memberLikesSet.add(suggestionId);
+
+    const suggestion: Suggestion | undefined = hangoutSuggestionState.suggestions.find((suggestion: Suggestion) => suggestion.suggestion_id === suggestionId);
+    suggestion && suggestion.likes_count++;
+
+    sortHangoutSuggestions();
+    displaySuggestionLike(suggestionElement);
+
+  } catch (err: unknown) {
+    console.log(err);
+    suggestionElement.classList.remove('like-pending');
+
+    if (!axios.isAxiosError(err)) {
+      popup('Failed to like suggestion.', 'error');
+      return;
+    };
+
+    const axiosError: AxiosError<AxiosErrorResponseData> = err;
+
+    if (!axiosError.status || !axiosError.response) {
+      popup('Failed to like suggestion.', 'error');
+      return;
+    };
+
+    const status: number = axiosError.status;
+    const errMessage: string = axiosError.response.data.message;
+    const errReason: string | undefined = axiosError.response.data.reason;
+
+    if (status === 400) {
+      popup('Failed to like suggestion.', 'error');
+      return;
+    };
+
+    popup(errMessage, 'error');
+
+    if (status === 409) {
+      displaySuggestionLike(suggestionElement);
+      return;
+    };
+
+    if (status === 401) {
+      if (errReason === 'authSessionExpired') {
+        handleAuthSessionExpired();
+        return;
+      };
+
+      if (errReason === 'notHangoutMember') {
+        popup(errMessage, 'error');
+        LoadingModal.display();
+
+        setTimeout(() => {
+          sessionStorage.removeItem('latestHangoutSection');
+          window.location.href = 'home';
+        }, 1000);
+      };
+    };
+  };
+};
+
+function displaySuggestionLike(suggestionElement: HTMLDivElement): void {
+  const suggestionLikesCountSpan: HTMLSpanElement | null = suggestionElement.querySelector('.rating-container span');
+
+  if (!suggestionLikesCountSpan) {
+    return;
+  };
+
+  const suggestionLikesCountString: string | null = suggestionLikesCountSpan.textContent;
+
+  if (!suggestionLikesCountString || !Number.isInteger(+suggestionLikesCountString)) {
+    return;
+  };
+
+  const suggestionLikesCount: number = +suggestionLikesCountString;
+  suggestionLikesCountSpan.textContent = `${suggestionLikesCount + 1}`;
+
+  suggestionElement.classList.remove('like-pending');
+  suggestionElement.classList.add('liked');
+};
+
+function removeSuggestionLike(suggestionElement: HTMLDivElement): void {
+  const suggestionLikesCountSpan: HTMLSpanElement | null = suggestionElement.querySelector('.rating-container span');
+
+  if (!suggestionLikesCountSpan) {
+    return;
+  };
+
+  const suggestionLikesCountString: string | null = suggestionLikesCountSpan.textContent;
+
+  if (!suggestionLikesCountString || !Number.isInteger(+suggestionLikesCountString)) {
+    return;
+  };
+
+  const suggestionLikesCount: number = +suggestionLikesCountString;
+  suggestionLikesCountSpan.textContent = `${(suggestionLikesCount - 1) < 0 ? 0 : suggestionLikesCount - 1}`;
+
+  suggestionElement.classList.remove('like-pending', 'liked');
+};
+
+export function sortHangoutSuggestions(): void {
+  const sortMode: 'likes' | 'votes' = hangoutSuggestionState.sortingMode
+  hangoutSuggestionState.suggestions.sort((a, b) => a[`${sortMode}_count`] - b[`${sortMode}_count`]);
 };